X11: Add window hint "GLFW_X11_BYPASS_FULLSCREEN"
This specifies whether to disable compositing of a
window in order to improve performance of rendering fullscreen.

We have to specify `GLFW_FALSE` when using pop-up windows or IME
on a fullscreen.

I don't know how much performance is actually improved by having
this turned on. If it doesn't affect performance much, then it
may be turned off by default.