Support soft fullscreen mode for each platform
Add a new window hint: `GLFW_SOFT_FULLSCREEN`.
This is especially necessary for the IME(Input Method Editor/Engine) to display
properly on the fullscreen window.

**Win32**

By default, the fullscreen window is a strong exclusive exceeding HWND_TOPMOST
and it doen't display OS notifications such as the sound volumn changed.
This hint enables to display other views poping up on the fullscreen window.
This can also effect transparency to fullscreen.

**macOS**

By default, the fullscreen window is exclusive.
This hint uses macOS's default fullscreen (the behavior of the green button
on the window's header).

We don't need `monitor` in this mode, so it is not used.
Specifying `monitor` in `glfwCreateWindow` makes window fullscreened,
but the value is not used inside.
The args of `glfwSetWindowMonitor` are not used except for `window` and
the function just toggles the fullscreened state.

These specifications have been made so that we can use this mode as
similar to the normal default mode as possible.

**X11**

This hint manages XProperty: `NET_WM_BYPASS_COMPOSITOR`.
By default, the compositing of the window is disabled to improve rendering performance.
However, this can cause problems with other views poping up on the window, such as IME.
This hint enables the compositing and the views poping up on the fullscreen to be
displayed correctly.