macOS: Add soft-fullscreen mode for apps using IME
This uses macOS's default fullscreen (the behavior of the green button
on the window's header).

This fullscreen is not exclusive and can display IME's candidate window.

We don't need `monitor` in this mode, so it is not used.
Specifying `monitor` in `glfwCreateWindow` makes window fullscreened,
but the value is not used inside.
The args of `glfwSetWindowMonitor` are not used except for `window` and
the function just toggles the fullscreened state.

These specifications have been made so that we can use this mode as
similar to the normal mode as possible.

Currently, this behavior can be switched by `SOFT_FULLSCREEN_MODE` value.

TODO

- Make `SOFT_FULLSCREEN_MODE` WindowHint.