Support soft fullscreen mode for each platform
Add a new window hint: `GLFW_SOFT_FULLSCREEN`.
This is especially necessary for the IME(Input Method Editor/Engine) to display
properly on the fullscreen window.

**Win32**

By default, the fullscreen window is a strong exclusive exceeding HWND_TOPMOST
and it doen't display OS notifications such as the sound volumn changed.
This hint enables to display other views poping up on the fullscreen window.
This can also effect transparency to fullscreen.

**macOS**

By default, the fullscreen window is so exclusive that it doesn't display IME.

This hint uses `NSApplication::setPresentationOptions` for fullscreen instead
of `NSWindow::setLevel`.
This allows us to create a fullscreen that can display IME.

`NSWindow::setLevel` is still used for fullscreen when this hint is off.
This is to avoid impact on existing behavior.
Possibly, we don't need to use it and we can use only
`NSApplication::setPresentationOptions`.

https://developer.apple.com/documentation/appkit/nsapplicationpresentationoptions?language=objc

**X11**

This hint manages XProperty: `NET_WM_BYPASS_COMPOSITOR`.
By default, the compositing of the window is disabled to improve rendering performance.
However, this can cause problems with other views poping up on the window, such as IME.
This hint enables the compositing and the views poping up on the fullscreen to be
displayed correctly.